Refactor Models.cs into logical files
Split 479-line Models.cs into smaller files:
- Models.cs (70 lines) - Core types, enums
- Models.Symbols.cs (159 lines) - Symbol/reference/member DTOs
- Models.Methods.cs (57 lines) - Method body/update DTOs
- Models.Diagnostics.cs (85 lines) - Diagnostic DTOs
- Models.CodeFix.cs (114 lines) - Code fix/rename DTOs

All files now under 300-line limit.

Closes #1
